Refactoring for new pipeline system
