+++ MVCForum.Core/Interfaces/Services/IEmailService.cs
@@ -13,5 +13,6 @@
         void ProcessMail(List<Email> emails);
         string EmailTemplate(string to, string content);
         string EmailTemplate(string to, string content, Settings settings);
+        void SendEmailConfirmationEmail(MembershipUser userToSave, bool manuallyAuthoriseMembers, bool memberEmailAuthorisationNeeded);
     }
 }
