+++ MVCForum.Website/Application/AppHelpers.cs
@@ -15,8 +15,8 @@
     using Core.Interfaces.Services;
     using Core.Models.Entities;
     using Core.Models.General;
+    using Core.Providers.Storage;
     using Core.Utilities;
-    using StorageProviders;
 
     public static class AppHelpers
     {
@@ -250,23 +250,6 @@
             return imageFileTypes.Any(file.Contains);
         }
 
-        public static Image GetImageFromExternalUrl(string url)
-        {
-            var httpWebRequest = (HttpWebRequest) WebRequest.Create(url);
-
-            using (var httpWebReponse = (HttpWebResponse) httpWebRequest.GetResponse())
-            {
-                using (var stream = httpWebReponse.GetResponseStream())
-                {
-                    if (stream != null)
-                    {
-                        return Image.FromStream(stream);
-                    }
-                }
-            }
-            return null;
-        }
-
         public static string MemberImage(string avatar, string email, Guid userId, int size)
         {
             if (!string.IsNullOrWhiteSpace(avatar))
@@ -290,132 +273,6 @@
             }
             //TODO - Return default image for category
             return null;
-        }
-
-        public static UploadFileResult UploadFile(HttpPostedFileBase file, string uploadFolderPath,
-            ILocalizationService localizationService, bool onlyImages = false)
-        {
-            var upResult = new UploadFileResult {UploadSuccessful = true};
-            const string imageExtensions = "jpg,jpeg,png,gif";
-            var fileName = Path.GetFileName(file.FileName);
-            var storageProvider = StorageProvider.Current;
-
-            if (fileName != null)
-            {
-                // Lower case
-                fileName = fileName.ToLower();
-
-                // Get the file extension
-                var fileExtension = Path.GetExtension(fileName);
-
-                //Before we do anything, check file size
-                if (file.ContentLength > Convert.ToInt32(SiteConstants.Instance.FileUploadMaximumFileSizeInBytes))
-                {
-                    //File is too big
-                    upResult.UploadSuccessful = false;
-                    upResult.ErrorMessage = localizationService.GetResourceString("Post.UploadFileTooBig");
-                    return upResult;
-                }
-
-                // now check allowed extensions
-                var allowedFileExtensions = SiteConstants.Instance.FileUploadAllowedExtensions;
-
-                if (onlyImages)
-                {
-                    allowedFileExtensions = imageExtensions;
-                }
-
-                if (!string.IsNullOrWhiteSpace(allowedFileExtensions))
-                {
-                    // Turn into a list and strip unwanted commas as we don't trust users!
-                    var allowedFileExtensionsList = allowedFileExtensions.ToLower().Trim()
-                        .TrimStart(',').TrimEnd(',').Split(',').ToList();
-
-                    // If can't work out extension then just error
-                    if (string.IsNullOrWhiteSpace(fileExtension))
-                    {
-                        upResult.UploadSuccessful = false;
-                        upResult.ErrorMessage = localizationService.GetResourceString("Errors.GenericMessage");
-                        return upResult;
-                    }
-
-                    // Remove the dot then check against the extensions in the web.config settings
-                    fileExtension = fileExtension.TrimStart('.');
-                    if (!allowedFileExtensionsList.Contains(fileExtension))
-                    {
-                        upResult.UploadSuccessful = false;
-                        upResult.ErrorMessage = localizationService.GetResourceString("Post.UploadBannedFileExtension");
-                        return upResult;
-                    }
-                }
-
-                // Store these here as we may change the values within the image manipulation
-                var newFileName = string.Empty;
-                var path = string.Empty;
-
-                if (imageExtensions.Split(',').ToList().Contains(fileExtension))
-                {
-                    // Rotate image if wrong want around
-                    using (var sourceimage = Image.FromStream(file.InputStream))
-                    {
-                        if (sourceimage.PropertyIdList.Contains(0x0112))
-                        {
-                            int rotationValue = sourceimage.GetPropertyItem(0x0112).Value[0];
-                            switch (rotationValue)
-                            {
-                                case 1: // landscape, do nothing
-                                    break;
-
-                                case 8: // rotated 90 right
-                                    // de-rotate:
-                                    sourceimage.RotateFlip(RotateFlipType.Rotate270FlipNone);
-                                    break;
-
-                                case 3: // bottoms up
-                                    sourceimage.RotateFlip(RotateFlipType.Rotate180FlipNone);
-                                    break;
-
-                                case 6: // rotated 90 left
-                                    sourceimage.RotateFlip(RotateFlipType.Rotate90FlipNone);
-                                    break;
-                            }
-                        }
-
-                        using (var stream = new MemoryStream())
-                        {
-                            // Save the image as a Jpeg only
-                            sourceimage.Save(stream, ImageFormat.Jpeg);
-                            stream.Position = 0;
-
-                            // Change the extension to jpg as that's what we are saving it as
-                            fileName = fileName.Replace(fileExtension, "");
-                            fileName = string.Concat(fileName, "jpg");
-                            file = new MemoryFile(stream, "image/jpeg", fileName);
-
-                            // Sort the file name
-                            newFileName = CreateNewFileName(fileName);
-
-                            // Get the storage provider and save file
-                            upResult.UploadedFileUrl = storageProvider.SaveAs(uploadFolderPath, newFileName, file);
-                        }
-                    }
-                }
-                else
-                {
-                    // Sort the file name
-                    newFileName = CreateNewFileName(fileName);
-                    upResult.UploadedFileUrl = storageProvider.SaveAs(uploadFolderPath, newFileName, file);
-                }
-
-                upResult.UploadedFileName = newFileName;
-            }
-
-            return upResult;
-        }
-
-        private static string CreateNewFileName(string fileName)
-        {
-            return $"{GuidComb.GenerateComb()}_{fileName.Trim(' ').Replace("_", "-").Replace(" ", "-").ToLower()}";
         }
 
         #endregion
